pyproject.toml
[project]
name = "shoprag"
version = "0.1.0"
description = "Retail RAG Chatbot for Amazon Reviews"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.3.5",
    "datasets>=2.19.0,<3.0.0",
    "pandas>=2.3.3",
    "sentence-transformers>=5.1.2",
    "tqdm>=4.67.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["backend"]
